requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.0.0
prompt_toolkit>=3.0.0
